Build for Android using buildozer
Bare-bones Android support added but it is not usable.  For some
reason it seems to hang whenever you try to press a key or even
connect the Twiddler via Bluetooth.

I think this one's dead in the water given how long Python 3
on Android support has been experimental for Kivy.  If I pick this
project up again it will likely be as a Kotlin + libGDX rewrite.